LIBILS-386. Moved "processor" classes into "dwetl/processor" folder
For consistency with the reader/writer classes, renamed "ProcessingStep"
class to "Processor", and moved into "dwetl/processor" folder.

Renamed "IdentityProcessing" to "IdentityProcessor"

Updated and renamed associated test classes, moving them into a
"test/processor" folder.

https://issues.umd.edu/browse/LIBILS-386
