+++ dwetl/processor/processor.py
@@ -1,4 +1,4 @@
-class ProcessingStep:
+class Processor:
     """
     Abstract class that encapsulates a processing step.
